requirements.txt
gradio==4.44.1
langchain==0.2.16
langchain-openai==0.3.22
chromadb==1.0.12
sentence-transformers==2.7.0
huggingface_hub>=0.30.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
requests==2.31.0
httpx==0.27.0